streamlit/consultancy/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
altair>=5.0.0
plotly>=5.18.0
openpyxl>=3.1.0
gspread>=5.12.0
google-auth>=2.23.0
python-dateutil>=2.8.0
